+++ common/src/main/java/immersive_aircraft/Items.java
@@ -25,6 +25,7 @@ public interface Items {
     Supplier<Item> AIRSHIP = register("airship", () -> new AircraftItem(baseProps().maxCount(1), (world) -> new AirshipEntity(Entities.AIRSHIP.get(), world)));
     Supplier<Item> BIPLANE = register("biplane", () -> new AircraftItem(baseProps().maxCount(1), (world) -> new BiplaneEntity(Entities.BIPLANE.get(), world)));
     Supplier<Item> GYRODYNE = register("gyrodyne", () -> new AircraftItem(baseProps().maxCount(1), (world) -> new GyrodyneEntity(Entities.GYRODYNE.get(), world)));
+    Supplier<Item> QUADROCOPTER = register("quadrocopter", () -> new AircraftItem(baseProps().maxCount(1), (world) -> new GyrodyneEntity(Entities.QUADROCOPTER.get(), world)));
 
     static Supplier<Item> register(String name, Supplier<Item> item) {
         Supplier<Item> register = Registration.register(Registries.ITEM, Main.locate(name), item);