+++ common/src/main/java/immersive_aircraft/mixin/KeyBindingMixin.java
@@ -11,50 +11,42 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
 /**
- * This injection hack allows multiple assignments to keys
+ * This injection hack allows multiple assignments to my keys
  */
-@Mixin(KeyBinding.class)
+@Mixin(value = KeyBinding.class, priority = 100)
 public class KeyBindingMixin {
     @Shadow
     @Final
     private static Map<String, KeyBinding> KEYS_BY_ID;
-    private static final Map<InputUtil.Key, List<KeyBinding>> immersiveAircraft$MULTI_KEY_TO_BINDINGS = new HashMap<>();
 
-
-    @Inject(method = "<init>(Ljava/lang/String;Lnet/minecraft/client/util/InputUtil$Type;ILjava/lang/String;)V", at = @At("TAIL"))
-    private void immersiveAircraft$init(String translationKey, InputUtil.Type type, int code, String category, CallbackInfo ci) {
-        immersiveAircraft$addKeyBinding((KeyBinding)(Object)this);
-    }
-
-    @Inject(method = "onKeyPressed(Lnet/minecraft/client/util/InputUtil$Key;)V", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "onKeyPressed(Lnet/minecraft/client/util/InputUtil$Key;)V", at = @At("HEAD"))
     private static void immersiveAircraft$onKeyPressed(InputUtil.Key key, CallbackInfo ci) {
-        List<KeyBinding> keyBinding = immersiveAircraft$MULTI_KEY_TO_BINDINGS.get(key);
+        List<MultiKeyBinding> keyBinding = MultiKeyBinding.KEY_TO_BINDING.get(key);
         if (keyBinding != null) {
             keyBinding.forEach(v -> ((KeyBindingAccessorMixin)v).setTimesPressed(((KeyBindingAccessorMixin)v).getTimesPressed() + 1));
         }
-        ci.cancel();
     }
 
-    @Inject(method = "setKeyPressed(Lnet/minecraft/client/util/InputUtil$Key;Z)V", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "setKeyPressed(Lnet/minecraft/client/util/InputUtil$Key;Z)V", at = @At("HEAD"))
     private static void immersiveAircraft$setKeyPressed(InputUtil.Key key, boolean pressed, CallbackInfo ci) {
-        List<KeyBinding> keyBinding = immersiveAircraft$MULTI_KEY_TO_BINDINGS.get(key);
+        List<MultiKeyBinding> keyBinding = MultiKeyBinding.KEY_TO_BINDING.get(key);
         if (keyBinding != null) {
             keyBinding.forEach(v -> v.setPressed(pressed));
         }
-        ci.cancel();
     }
 
     @Inject(method = "updateKeysByCode()V", at = @At("HEAD"))
     private static void immersiveAircraft$updateKeysByCode(CallbackInfo ci) {
-        immersiveAircraft$MULTI_KEY_TO_BINDINGS.clear();
+        MultiKeyBinding.KEY_TO_BINDING.clear();
         for (KeyBinding keyBinding : KEYS_BY_ID.values()) {
-            immersiveAircraft$addKeyBinding(keyBinding);
+            if (keyBinding instanceof MultiKeyBinding kb) {
+                MultiKeyBinding.KEY_TO_BINDING.computeIfAbsent(kb.customBoundKey, v -> new LinkedList<>()).add(kb);
+            }
         }
     }
 
@@ -64,9 +56,5 @@ public class KeyBindingMixin {
         if (((Object)this) instanceof MultiKeyBinding || other instanceof MultiKeyBinding) {
             cir.setReturnValue(((Object)this) == other);
         }
-    }
-
-    private static void immersiveAircraft$addKeyBinding(KeyBinding keyBinding) {
-        immersiveAircraft$MULTI_KEY_TO_BINDINGS.computeIfAbsent(((KeyBindingAccessorMixin)keyBinding).getBoundKey(), (k) -> new LinkedList<>()).add(keyBinding);
     }
 }