+++ common/src/main/java/immersive_aircraft/network/c2s/CommandMessage.java
@@ -23,6 +23,7 @@ public class CommandMessage implements Message {
         if (e.getRootVehicle() instanceof AircraftEntity aircraft) {
             if (key == Key.DISMOUNT) {
                 e.stopRiding();
+                aircraft.chill();
                 aircraft.setVelocity(fx, fy, fz);
             }
         }