+++ common/src/main/java/immersive_aircraft/util/obj/VertexNormal.java
@@ -1,4 +1,4 @@
-package immersive_aircraft.util.obj.oobjloader;
+package immersive_aircraft.util.obj;
 
 // This code was written by myself, Sean R. Owens, sean at guild dot net,
 // and is released to the public domain. Share and enjoy. Since some
@@ -9,12 +9,18 @@ package immersive_aircraft.util.obj.oobjloader;
 // In addition this code may also be used under the "unlicense" described
 // at http://unlicense.org/ .  See the file UNLICENSE in the repo.
 
-public class VertexPosition {
-    public final float x;
-    public final float y;
-    public final float z;
+public class VertexNormal {
+    public float x;
+    public float y;
+    public float z;
 
-    public VertexPosition(float x, float y, float z) {
+    public void add(float x, float y, float z) {
+        this.x += x;
+        this.y += y;
+        this.z += z;
+    }
+
+    public VertexNormal(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;