+++ fabric/src/main/java/immersive_aircraft/fabric/CommonFabric.java
@@ -3,10 +3,7 @@ package immersive_aircraft.fabric;
 import immersive_aircraft.*;
 import immersive_aircraft.fabric.cobalt.network.NetworkHandlerImpl;
 import immersive_aircraft.fabric.cobalt.registration.RegistrationImpl;
-import immersive_aircraft.server.Command;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
-import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 
 public final class CommonFabric implements ModInitializer {
     @Override
@@ -18,8 +15,6 @@ public final class CommonFabric implements ModInitializer {
         Sounds.bootstrap();
         Entities.bootstrap();
         Messages.bootstrap();
-
-        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> Command.register(dispatcher));
 
         ItemGroups.GROUP = FabricItemGroup.builder(ItemGroups.getIdentifier())
                 .displayName(ItemGroups.getDisplayName())