+++ changelog.md
@@ -1,3 +1,17 @@
+# 0.3.0
+
+* Added Quadrocopter
+* Fixed non-pilot dismounting
+* Switched to less intrusive multi keybindings injection to restore mod compatibility
+* Fixed controller animation of airship
+* Added support for modmenu and cloth config
+* Added item tooltip description
+* You can no longer apply knockback to your vehicle
+* Fixed the Biplane phantom push when leaving
+* Fixed first person camera offset
+* Vehicles can have a separate camera perspective, and is third person by default (configurable)
+* Creative mode destroys airplanes immediately
+
 # 0.2.1
 
 * Fixed Forge on 1.19.x